Constructors accept flattening.  Geod does not report m12 by default.
Constructors accept f which is interpreted as flattening if f < 1;
otherwise set flattening to 1/f.  This is a compatible change for
spheres and oblate ellipsoids.  For prolate ellipsoids this in an
incompatible change.

Similar changes made to utility and matlab interface.

Most classes now have a Flattening() inspector function (in addition to
InverseFlattening()).

Geod -f now reports m12 M12 M21 and S12.  m12 is not reported by
default.