Allow use of a multiprecision type (boost's float128) for Math::real. Issues:
(1) GeodesicExact.cpp takes ages to compile (a ccache issue?).
(2) setprecision(0) is interpreted as print all the digits (so formatted
output is wrong in some case and one of the tests is broken) (NOT
fixed).
(3) volatile declarations don't work (fixed with a macro).
(4) Deal with the str<T> issue with isfinite with a Math::real overload
(5) readarray optimization disabled for multiprec
(6) Needed to get rid of mixed precision (extended + real) in some test
programs.