Revert URL for random doc; get rid of deadwood in GeodesicExact.cpp.
Revert URL, e.g.,

  https://geographiclib.sourceforge.io/misc/geod-addenda.html

back to

  https://geographiclib.sourceforge.io/misc/geod-addenda.html

The reason is that the final position might change again -- this will
all be dealt with in geographiclib-doc.

Remove unused 24th + 27th order series from GeodesicExact.cpp.  We can
now experiment with 64th order series, but this is very large and
definitely shouldn't be included in the distribution.