Make Trigfun::root more robust
Insert bisection if dx doesn't decrease by a factor of 2.  Tighten up
vtol exit criterion.  This addresses convergence failures with PREC=5
and maxcnt = 150.

Also use tol = 0 as the default value in these functions and substitute
epsilon() if tol is 0.

Insert an indicator argument to allow debugging of recursive calls to
this function.