Standardize on oblpro = false, merid = true treatment
Eliminate as much special treatment of biaxial cases as possible.  Now
only meridional geodesics on biaxial ellipsoids get special treatment;
this is unavoidable because the nature of the singularity in the f
functions changes in the biaxial limit.

For gamma != 0, the biaxial cases are treated the same as for triaxial
ellipsoids.

For gamma == 0, the geodesic is broken into segments deliminated by
passage through the pole (biaxial) or umbilic (triaxial).