Start hybridalt implementation.
This refers to solving the hybrid problem with a target longitude
instead of latitude in an attempt to reduce the errors in the inverse
problem for prolate and near prolate ellipsoids.

Initial try was to have Triaxial::findroot use this modified problem.
But this quickly runs into problems on prolate ellipsoid where a
geodesic might wrap around the ellipsoid multiple times to reach the
target longitude.

It turns out that  Triaxial::findroot finds an accurate alp1 using the
original target latitude hybird solution.  The error in the inverse
problem occurs in the final

    d = lf.Hybrid(fic, bet2, bet2a, omg2a, alp2)

to obtain the parameters needed to compute the distance.  If alp2 is
close to +/-90, then the geodesic intersects bet = bet2 at a grazing
angle.  Better to find the intersection omg = omg2 with

    d = lf.Hybrid(fic, betp ? bet2 : omg2, bet2a, omg2a, alp2, betp);

and choose betp based on alp2.

With --hybridalt, we get a big improvement in the accuracy of s12:

                   orig                  hybridalt
                mean    max             mean    max
  testhu        2       348             2       84
  testobl       2       1144            2       82
  testoblx      2       268             2       36
  testphu       35      151752          3       2084
  testpro       1844    14561760        3       88
  testprox      12      114614          4       288
  testseta      2       6388            2       82
  testset       4       20916           2       88
  testspha      2       239             2       94
  testsphb      2       720             2       90
  testsphc      3       1410            2       88
  testsphd      2058    12622306        2       82

Dangling issues

(1) The betp == false requires omg1.s() <= fabs(omg2.s()).  I.e., omg1
must if as close or closer to the omg = 0/180 pole than omg2.  (This
matches the requirement on bet1/bet2 which is obtained by swap12.)  It's
not clear what goes wrong if this condition is not met, since alp1 will
still allow the geodesic to intersect the line omg = omg2.  Investigate
what happens.

(2) There's something amiss in fline::Hybrid.  Origial code includes

  tht2b.reflect(false, fic.Ex < 0);

But this reflection causes

  echo -88 21 88 -111 | ./Geod3Solve -i $SET --hybridalt

to give tau12 = 0 instead of 111+21 = 132.  However removing the reflection
(fic.Ex < 0 -> false) causes problems for other cases.  In the meantime, use

  tht2b.reflect(false, betp && fic.Ex < 0);

and this is ugly but "works".  But still need to track down what's going
on here.