Fix alp2 rebasing, etc.
Add direct.m to check monotonicity (increasing) property for theta +
psi.  Uncovers problem with rebasing of alp2; this needs to be done
after the alp2a.reflect(false, false, true).  This also uncovered a
problem with Angle::rebase because of the ambiguty in the representions
of +/- 180d.  Fix by intruding Angle::n0 and Angle::setno.

Remove some dead code (treatment of kxp2() == 0 in the else clause of
if (kxp2() == 0)).