Implement new2d for the direct problem.
This uses 2d Newton's method to solve the direct problem (as opposed to
nested Newton's method).  This uses bounds on the solution for (x,y) and
a bisection step if Newton's method gives a solution outside the
bounding box.

This method is flawed because the bounding box doesn't reliably get
smaller by a factor of 2 or 4 on each bisection step.

We can do much better than this.