Initial implementation of newt2d for umbilic geodesics.
Running into several instances of failures to converge.  Bandaids are:

  tolmult (allowed Newton iterations to lie outside the bounding box)
  enlarging the initial bounding box (shouldn't be needed)

Tentatively ascribe the problems to non-monotonicity of the fz, gz
functions.

To do:

* mainly to rule this out ... check the initial bounding box (this
will involved evaluating f/g on the midpoints of the edges);

* check that the fz and gz functions are non-decreasing.

If the second is the problem, then modify zvals class to replace
non-conforming fz and gz values by clamped values.