Fix long double issues with prolate ellipsoids
Fixed bugs elliptocart2 for biaxial poles using sign of cos(beta) for
oblate and sin(omega) for prolate.  Same fix also done for Octave
implementation.

But this isn't enough.  The problem (I think) is that the sign of 0
isn't properly maintained when cardinal directions are added to the
angle.  Fix by adjusting previous cos(phi) = max(cos(phi), 0) fix to set
cos(phi) to eps/2048 if signbit(cos(phi)).

Another issue to be investigated:

quad errors are suspiciously large.  See line 2276 of testset.txt.  This
is an umbilical geodesic, but the data in testset.txt isn't precise
enough to give gamma = 0 for the direct calculation.  So this is a
geodesic is tiny non-zero gamma.

Can recompute this line and print with higher precision giving (-p 13
and -p 20 results)
-90 180 -31.678276203968430803 -16 152 -34.632417851313538856 0.54646403823611170442 0.50489789511517110353 0.80569806066626949844 0.73963052819047320946
-90 180 -31.6782762039684308026562168 -16 152 -34.6324178513135388559304218 0.546464038236111704419875801 0.50489789511517110353 0.80569806066626949844 0.73963052819047320946

Probably can trigger similar problem at double precision by changing the
azimuths slightly.