Experiment using the secant method in newt2.
This was an attempt to reduce the number of iteration.  This doesn't
help; but the code is left in place.

Need to explore swapping omg1 <-> omg2 in the inverse problem.  Make
point 1 not only have bet1 closer to -90 but also have omg1 closer to 0.
Then the _hybridalt solution would no longer have to include the test
omg1.s() <= fabs(omg2.s()).