Introducing swapomg in Triaxial.
This get the hybridalt logic used more frequently (to do: remove the
test omg1.s() <= fabs(omg2.s()) in using hybridalt.  This is effective
in reducing the errors in the distance particularly for testphy (2084 ->
72) and testprox (288 -> 60).

At present this include a lame treatment of the azimuth conversion for
swapomg.  To do: improve this.