Fix bug in Geodesic with meridional lines on very prolate ellipsoids
For f < -2 (b > 3*a), the meridional line was sometimes incorrectly
returned as the shortest geodesic.

This was caused by assuming that sig12 < 1 guarantees the meridional
geodesic is shortest (even though m12 < 0).  (This was added to deal
with the possibility that m12 < 0 for very short geodesics.)  The
inverse calculation

  echo 30.61 0 30.61 180 | GeodSolve -i -e 1e6 -2.1 -E
   0.00000000 180.00000000 12745815.681 # incorrect
  11.61546928 168.38453072 12742487.911 # correct

exhibits the bug.  The bug was present in the series and the exact
implementations (Geodesic and GeodesicExact).  Of course, the series
implementation is, in any case, returning inaccurate results for f < -2.

Tests GeodSolve100 + 101 added to check this.